+++ Carbon.FileSystem/Carbon.FileSystem.psm1
@@ -15,48 +15,21 @@
 #Requires -Version 5.1
 Set-StrictMode -Version 'Latest'
 
-# Functions should use $moduleRoot as the relative root from which to find
-# things. A published module has its function appended to this file, while a 
+# Functions should use $script:moduleRoot as the relative root from which to find
+# things. A published module has its function appended to this file, while a
 # module in development has its functions in the Functions directory.
-$moduleRoot = $PSScriptRoot
+$script:moduleRoot = $PSScriptRoot
+$modulesRoot = Join-Path -Path $script:moduleRoot -ChildPath 'Modules'
 
-Add-Type @'
-using System;
-using System.Text;
-using System.Collections.Generic;
-using System.Runtime.InteropServices;
+Import-Module -Name (Join-Path -Path $modulesRoot -ChildPath 'PureInvoke' -Resolve) `
+              -Function @('Invoke-KernelFindFileName')
 
-namespace Carbon.FileSystem
-{
-  public static class Kernel32
-  {
-
-    #region WinAPI P/Invoke declarations
-    [DllImport("kernel32.dll", CharSet = CharSet.Unicode, SetLastError = true)]
-    public static extern IntPtr FindFirstFileNameW(string lpFileName, uint dwFlags, ref uint StringLength, StringBuilder LinkName);
-
-    [DllImport("kernel32.dll", CharSet = CharSet.Unicode, SetLastError = true)]
-    public static extern bool FindNextFileNameW(IntPtr hFindStream, ref uint StringLength, StringBuilder LinkName);
-
-    [DllImport("kernel32.dll", SetLastError = true)]
-    public static extern bool FindClose(IntPtr hFindFile);
-
-    [DllImport("kernel32.dll", CharSet = CharSet.Unicode, SetLastError = true)]
-    public static extern bool GetVolumePathName(string lpszFileName, [Out] StringBuilder lpszVolumePathName, uint cchBufferLength);
-
-    public static readonly IntPtr INVALID_HANDLE_VALUE = (IntPtr)(-1); // 0xffffffff;
-    public const int MAX_PATH = 65535; // Max. NTFS path length.
-    #endregion
-   }
-}
-'@
-
-# Store each of your module's functions in its own file in the Functions 
-# directory. On the build server, your module's functions will be appended to 
+# Store each of your module's functions in its own file in the Functions
+# directory. On the build server, your module's functions will be appended to
 # this file, so only dot-source files that exist on the file system. This allows
 # developers to work on a module without having to build it first. Grab all the
 # functions that are in their own files.
-$functionsPath = Join-Path -Path $moduleRoot -ChildPath 'Functions\*.ps1'
+$functionsPath = Join-Path -Path $script:moduleRoot -ChildPath 'Functions\*.ps1'
 if( (Test-Path -Path $functionsPath) )
 {
     foreach( $functionPath in (Get-Item $functionsPath) )